sqlite_protobuf: Reuse Message memory
Summary:
This patch adds reuse of the Message object's memory to reduce the
memory management overhead.  Using Arenas was not helpful and in fact
slowed things down slightly.  The patch speeds up the script of Dxxxx
to from 3.5s to 2.2s and the script of Dxxxx from 7.3s to 6.7s.

Ref Txxxx

Depends on Dxxxx

Test Plan: ran the test scripts

Reviewers: wandrews, pkhuong, ahicks

Reviewed By: wandrews, pkhuong

Maniphest Tasks: Txxxx

Differential Revision: https://foo.backtrace.io/Dxxxx

Signed-off-by: Paul Khuong <pkhuong@backtrace.io>