sqlite_protobuf: Reuse Message memory for small messages
Summary:
We attempt to reuse Message objects as long as the message
data doesn't shrink too quickly compared to the maximum
message data size we have parsed.  This patch adds some
robustness to the reallocation logic to avoid reallocating
if some of the messages are really small: messages smaller than
a fixed constant are considered to be of that size when
checking for shrinking.

Ref Txxxx

Depends on Dxxxx

Test Plan: ran the scripts with temporary logging

Reviewers: pkhuong, wandrews, ahicks

Reviewed By: pkhuong

Maniphest Tasks: Txxxx

Differential Revision: https://foo.backtrace.io/Dxxxx

Signed-off-by: Paul Khuong <pkhuong@backtrace.io>