proto_table: Add `impl` to application specific row types.
Summary:
The `PROTO_RESULT_LIST` macro declares application specific
result list and row types that are memory layout compatible with
`proto_result_list` and `proto_result_row`, respectively, but
with the `proto` field of the row pointing to an application specific
type.

The new types are constructed so that the same field names
may be used as for `proto_result_list` and `proto_result_row`.
Additionally, a value of the new result list type may be accessed
as a `proto_result_list` using a union field `impl`.

This patch adds an `impl` union field to the new row type for
the client to be able to access the row as a `proto_result_row`.

Ref Txxxxx

Test Plan: build everything

Reviewers: pkhuong, wandrews, ahicks, jshufro

Reviewed By: pkhuong

Maniphest Tasks: Txxxxx

Differential Revision: https://foo.backtrace.io/Dxxxx